moved all the functions from main.py to plot.py to make the merge easier
